fixes qureis eslint rules
